# Conflicts:
#	src/main/java/pe/edu/pe/smartrent_backend/Controllers/ContractController.java
#	src/main/java/pe/edu/pe/smartrent_backend/Controllers/FavoriteController.java
#	src/main/java/pe/edu/pe/smartrent_backend/Controllers/UserBackgroundController.java
#	src/main/java/pe/edu/pe/smartrent_backend/Controllers/UserController.java
#	src/main/java/pe/edu/pe/smartrent_backend/DTOS/favoriteDTOS/FavoriteCompleteDTO.java
#	src/main/java/pe/edu/pe/smartrent_backend/DTOS/favoriteDTOS/FavoriteDTO.java
#	src/main/java/pe/edu/pe/smartrent_backend/DTOS/userDTOS/UserEnabledByRoleDTO.java
#	src/main/java/pe/edu/pe/smartrent_backend/DTOS/userbackgorundDTOS/UserBackgroundMonthlyDTO.java
#	src/main/java/pe/edu/pe/smartrent_backend/DTOS/userbackgorundDTOS/UserBackgroundSourceDTO.java
#	src/main/java/pe/edu/pe/smartrent_backend/DTOS/userbackgorundDTOS/UserBackgroundTypeFrequencyDTO.java
#	src/main/java/pe/edu/pe/smartrent_backend/Entities/Roles.java
#	src/main/java/pe/edu/pe/smartrent_backend/Repositories/IContractRepository.java
#	src/main/java/pe/edu/pe/smartrent_backend/Repositories/IRiskReportRepository.java
#	src/main/java/pe/edu/pe/smartrent_backend/Repositories/IUserBackgroundRepository.java
#	src/main/java/pe/edu/pe/smartrent_backend/Repositories/IUserRepository.java
#	src/main/java/pe/edu/pe/smartrent_backend/ServicesImplements/ContractServiceImplement.java
#	src/main/java/pe/edu/pe/smartrent_backend/ServicesImplements/FavoriteServiceImplements.java
#	src/main/java/pe/edu/pe/smartrent_backend/ServicesImplements/UserBackgroundServicesImplements.java
#	src/main/java/pe/edu/pe/smartrent_backend/ServicesImplements/UserServicesImplements.java
#	src/main/java/pe/edu/pe/smartrent_backend/ServicesInterfaces/IContractService.java
#	src/main/java/pe/edu/pe/smartrent_backend/ServicesInterfaces/IFavorite.java
#	src/main/java/pe/edu/pe/smartrent_backend/ServicesInterfaces/IUser.java
#	src/main/java/pe/edu/pe/smartrent_backend/ServicesInterfaces/IUserBackground.java

+++ src/main/java/pe/edu/pe/smartrent_backend/DTOS/userDTOS/UserEnabledByRoleDTO.java
@@ -2,8 +2,8 @@ package pe.edu.pe.smartrent_backend.DTOS.userDTOS;
 
 public class UserEnabledByRoleDTO {
     private String role;
-    private Long enabled;
-    private Long disabled;
+    private int enabled;
+    private int disabled;
 
     public String getRole() {
         return role;
@@ -13,19 +13,19 @@ public class UserEnabledByRoleDTO {
         this.role = role;
     }
 
-    public Long getEnabled() {
+    public int getEnabled() {
         return enabled;
     }
 
-    public void setEnabled(Long enabled) {
+    public void setEnabled(int enabled) {
         this.enabled = enabled;
     }
 
-    public Long getDisabled() {
+    public int getDisabled() {
         return disabled;
     }
 
-    public void setDisabled(Long disabled) {
+    public void setDisabled(int disabled) {
         this.disabled = disabled;
     }
 }
